+++ main/classes/TagProbe.h
@@ -5,6 +5,7 @@ class TagProbe{
 private:
 	int& method;
 	const char*& particleName;
+	const char*& canvasWatermark;
 	const char*& directoryToSave;
 	const char*& particleType;
 	InvariantMass& ObjMass;
@@ -12,11 +13,11 @@ private:
 public:
 	const char* tagOrProbe = NULL;
 
-	PtEtaPhi Pt  {this->method, this->particleName, this->directoryToSave, this->particleType, this->ObjMass, this->tagOrProbe,
+	PtEtaPhi Pt  {this->method, this->particleName, this->canvasWatermark, this->directoryToSave, this->particleType, this->ObjMass, this->tagOrProbe,
 		"Pt",  "p_{t}", "GeV/c", "Transversal Momentum", 100,  0.00, 100.00, 1};
-	PtEtaPhi Eta {this->method, this->particleName, this->directoryToSave, this->particleType, this->ObjMass, this->tagOrProbe,
+	PtEtaPhi Eta {this->method, this->particleName, this->canvasWatermark, this->directoryToSave, this->particleType, this->ObjMass, this->tagOrProbe,
 		"Eta", "#eta",  "", 	 "Pseudorapidity",       100, -2.50,   2.50};
-	PtEtaPhi Phi {this->method, this->particleName, this->directoryToSave, this->particleType, this->ObjMass, this->tagOrProbe,
+	PtEtaPhi Phi {this->method, this->particleName, this->canvasWatermark, this->directoryToSave, this->particleType, this->ObjMass, this->tagOrProbe,
 		"Phi", "#phi",  "rad",   "Azimuthal Angle",       79, -3.15,   3.15};
 
 	void subtractSigHistograms()
@@ -80,12 +81,14 @@ public:
 
 	TagProbe(int& method,
 		const char*& particleName,
+		const char*& canvasWatermark,
 		const char*& directoryToSave,
 	 	const char*& particleType,
 	 	InvariantMass& ObjMass,
 	 	const char*  tagOrProbe)
 		  : method(method),
 		    particleName(particleName),
+		    canvasWatermark(canvasWatermark),
 		    directoryToSave(directoryToSave),
 		    particleType(particleType),
 		    ObjMass(ObjMass),