+++ main/classes/Type.h
@@ -7,14 +7,15 @@ private:
 	int& method;
 	const char*& ressonance;
 	const char*& particleName;
+	const char*& canvasWatermark;
 	const char*& directoryToSave;
 
 public:
 	const char* particleType = NULL;
 
-	InvariantMass Mass  {this->method, this->ressonance, this->particleName, this->directoryToSave, this->particleType};
-	TagProbe      Tag   {this->method, this->particleName, this->directoryToSave, this->particleType, this->Mass, "Tag"};
-	TagProbe      Probe {this->method, this->particleName, this->directoryToSave, this->particleType, this->Mass, "Probe"};
+	InvariantMass Mass  {this->method, this->ressonance, this->particleName, this->canvasWatermark, this->directoryToSave, this->particleType};
+	TagProbe      Tag   {this->method, this->particleName, this->canvasWatermark, this->directoryToSave, this->particleType, this->Mass, "Tag"};
+	TagProbe      Probe {this->method, this->particleName, this->canvasWatermark, this->directoryToSave, this->particleType, this->Mass, "Probe"};
 
 	void defineMassHistogramNumbers(double xMin, double xMax, int nBins, int decimals = 3)
 	{
@@ -105,11 +106,13 @@ public:
 	Type(int& method,
 		const char*& ressonance,
 		const char*& particleName,
+		const char*& canvasWatermark,
 		const char*& directoryToSave,
 	 	const char*  particleType)
 		  : method(method),
 		    ressonance(ressonance),
 		    particleName(particleName),
+		    canvasWatermark(canvasWatermark),
 		    directoryToSave(directoryToSave),
 		    particleType(particleType)
 	{}